Changed the directory of STACK&QUEUE folder to C++ folder
