+++ c++/STACKS&QUEUE/Readme.md
@@ -13,3 +13,4 @@
 3. Q3.cpp : SLIDING WINDOW MAXIMUM
 4. Q4.cpp : ROTTING ORANGES
    
+## THE DIRECTORY HS BEEN CHANGED FROM STACK&QUEUE TO C++
